back end api for explore page

+++ src/main/java/com/launchcode/liftoff/the/bugfest/club/data/DestinationRepository.java
@@ -1,9 +1,10 @@
 package com.launchcode.liftoff.the.bugfest.club.data;
 
 import com.launchcode.liftoff.the.bugfest.club.models.Destination;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DestinationRepository extends CrudRepository<Destination,Integer> {
+public interface DestinationRepository extends CrudRepository<Destination,Long> {
 }
