+++ src/main/java/com/launchcode/liftoff/the/bugfest/club/data/LocationRepository.java
@@ -1,10 +1,10 @@
 package com.launchcode.liftoff.the.bugfest.club.data;
 
-import com.launchcode.liftoff.the.bugfest.club.models.Destination;
 import com.launchcode.liftoff.the.bugfest.club.models.Location;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocationRepository extends CrudRepository<Location,Integer> {
+public interface LocationRepository extends CrudRepository<Location,Long> {
 }
